Put quotes around range of addresses for thp ranges
The code sets the huge_addr file in range mode by echoing the semicolon
seperated list of addresses ranges and piping that to tee which writes to
/sys/whatever/huge_addr. If there was more than one range, bash would
interpret the semicolon as the end of a command and interpret the rest of
the addresses as an invalid command. Therefore, huge_addr would not be
written.
This patch simply puts double quotes around the list of addresses to prevent
this from happening.